refactor(regeneration): move RelationshipHop/ReachedPath into core models
These describe a relationship chain -- kind, relationship identifier, direction
-- with nothing GraphQL about them, yet core imported them from the graphql
analyzer, inverting the layer dependency. Move them to the regeneration models
so the analyzer produces them from core and core no longer reaches up into
graphql.
